Refactor to avoid prop drilling
